backend/requirements.txt
# Backend Requirements
# Install with: pip install -r requirements.txt

flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.12.0
keras>=2.12.0
pillow>=9.5.0
opencv-python>=4.7.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
tqdm>=4.65.0
